+++ src/Azure.Functions.Sdk/release_notes.md
@@ -4,6 +4,6 @@
 - My change description (#PR/#issue)
 -->
 
-### Azure.Functions.Sdk 1.0.0
+### Azure.Functions.Sdk 1.0.1
 
-- Support `FunctionsStartupAttribute` scanning (#3482)
+- Fix ZipDeploy publishing on clean project trees (#3504)